Cast math operands before assignment
See https://rules.sonarsource.com/java/RSPEC-2184

+++ src/main/java/com/aoindustries/aoserv/daemon/email/SmtpRelayManager.java
@@ -374,6 +374,6 @@ public class SmtpRelayManager extends BuilderThread implements Runnable {
 
 	@Override
 	public long getProcessTimerMaximumTime() {
-		return (long)30*60*1000;
+		return 30L * 60 * 1000;
 	}
 }
